+++ src/utils.ts
@@ -1,14 +1,13 @@
+import * as path from 'path'
 import * as vscode from 'vscode'
 import * as compilerUtils from './compiler/utils'
+import * as ext from './extension'
 
-import { createHash } from 'crypto'
-import { workspace } from 'vscode'
-import { diagnosticsCollection } from './extension'
 import { intellisense } from './intellisense/features'
-import { VTSEnvStorage, bundleInfoRepository } from './intellisense/utils'
+import { bundleInfoRepository } from './intellisense/utils'
 
 export const getExtOption = <T>(option: string): T =>
-  workspace.getConfiguration('wglscript').get(option) as T
+  vscode.workspace.getConfiguration('wglscript').get(option) as T
 
 export function requestOpenWglScriptWorkspace() {
   vscode.window
@@ -20,17 +19,6 @@ export function requestOpenWglScriptWorkspace() {
       if (v !== 'Open Folder') return
       vscode.commands.executeCommand('workbench.action.files.openFolder')
     })
-}
-
-export function getHash(content: string) {
-  const hash = createHash('sha256')
-  hash.update(content)
-  return hash.digest('hex')
-}
-
-export function verifyHash(content: string, originalHash: string) {
-  const currentHash = getHash(content)
-  return currentHash === originalHash
 }
 
 export function arrayToMap<T>(array: Array<T>) {
@@ -59,11 +47,10 @@ export function logtime<T extends AnyFn>(fn: T, ...args: Parameters<T>): ReturnT
 
 export const restartService = () => {
   try {
-    diagnosticsCollection.clear()
-    VTSEnvStorage.clear()
+    ext.diagnosticsCollection.clear()
     bundleInfoRepository.clear()
-
     compilerUtils.astStorage.clear()
+
     compilerUtils.gls.code = ''
     compilerUtils.gls.sourcemap = ''
     compilerUtils.gls.modules = new Map()
@@ -197,5 +184,97 @@ export async function showModuleInfo() {
     console.log(error)
     vscode.window.showErrorMessage(`WGLToolchain ${error}`)
     restartService()
+  }
+}
+
+export let lastEdit: Date = new Date()
+
+export async function debaunceChangeTextDocument(signal: Date) {
+  await new Promise(r => setTimeout(r, 2000))
+  if (signal === lastEdit) return true
+}
+
+/**
+ * Handle document updates
+ * @param document
+ * @param force handle document updates regardless of {@link vscode.TextDocument.isDirty isDirty} TextDocument prop
+ * @returns
+ */
+export function didChangeTextDocumentHandler({
+  document,
+  force
+}: {
+  document: Pick<vscode.TextDocument, 'isDirty' | 'uri' | 'fileName' | 'version' | 'languageId'>
+  force?: boolean
+}) {
+  if (document.languageId !== 'javascript') return
+  if (force !== true && !document.isDirty) return
+
+  const activeTE = vscode.window.activeTextEditor
+
+  if (!activeTE) return
+
+  const wsPath = vscode.workspace.getWorkspaceFolder(activeTE.document.uri)?.uri.fsPath
+
+  if (!wsPath) return
+
+  lastEdit = new Date()
+
+  const normalized = compilerUtils.normalizePath(document.uri.fsPath, wsPath)
+
+  if (intellisense.modulesWithError.has(normalized)) {
+    intellisense.modulesWithError.delete(normalized)
+  }
+
+  for (const [entry, info] of bundleInfoRepository) {
+    if (info.dependencies.find(d => d === normalized)) {
+      // exit not implemented in tsvfs
+      // if (info.env) info.env.sys.exit(0)
+      bundleInfoRepository.delete(entry)
+    }
+  }
+
+  for (const [entry, deps] of intellisense.moduleReferencesStorage)
+    if (deps.find(d => d === normalized)) intellisense.moduleReferencesStorage.delete(entry)
+
+  if (compilerUtils.astStorage.has(normalized)) {
+    compilerUtils.astStorage.delete(normalized)
+  }
+
+  if (compilerUtils.gls.code !== '') {
+    if (mapToArray(compilerUtils.gls.modules).indexOf(normalized.toLowerCase()) !== -1) {
+      compilerUtils.gls.code = ''
+      compilerUtils.gls.sourcemap = ''
+      compilerUtils.gls.modules = new Map()
+    }
+  }
+
+  const diagnosticsStrategy = getExtOption<'onchange' | 'onsave' | 'disabled'>(
+    'intellisense.requestStrategy.diagnostics'
+  )
+
+  if (vscode.window.activeTextEditor && diagnosticsStrategy === 'onchange') {
+    const activeDoc = vscode.window.activeTextEditor.document
+    if (document.fileName === activeDoc.fileName) {
+      const wsPath = vscode.workspace.getWorkspaceFolder(activeDoc.uri)?.uri.fsPath
+
+      if (wsPath) {
+        if (vscode.window.activeTextEditor?.document.version !== document.version) return
+
+        debaunceChangeTextDocument(lastEdit).then(
+          passed =>
+            passed &&
+            intellisense
+              .getDiagnostics({ fileName: activeDoc.fileName }, wsPath)
+              .then(diagnostics => {
+                if (!diagnostics) return
+
+                ext.diagnosticsCollection.clear()
+                for (const [m, d] of diagnostics)
+                  ext.diagnosticsCollection.set(vscode.Uri.file(path.join(wsPath, m)), d)
+              })
+        )
+      }
+    }
   }
 }